Client code refactoring. Trying some GUI libraries.
