+++ client/build.gradle
@@ -4,4 +4,5 @@ mainClassName = 'ClientHelloMain'
 
 dependencies {
     implementation project(':core')
+    implementation 'com.jfoenix:jfoenix:8.0.4' // Java 8
 }
