+++ client/src/main/resources/openingLayout.fxml
@@ -5,10 +5,13 @@
 <?import javafx.scene.layout.Pane?>
 <?import javafx.scene.text.Font?>
 
-<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="300.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.121" xmlns:fx="http://javafx.com/fxml/1" fx:controller="it.cwmp.controller.OpeningController">
+<?import com.jfoenix.controls.JFXButton?>
+<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="300.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.121" xmlns:fx="http://javafx.com/fxml/1" fx:controller="it.cwmp.client.controller.OpeningController">
 
     <children>
-        <Button layoutX="111.0" layoutY="195.0" mnemonicParsing="false" onAction="#onClickSignIn" prefHeight="42.0" prefWidth="102.0" text="Login" />
+        <JFXButton buttonType="RAISED" ripplerFill="#de3a3a" layoutX="111.0" layoutY="195.0" mnemonicParsing="false" onAction="#onClickSignIn" prefHeight="42.0" prefWidth="102.0" text="Login" />
+        <JFXButton prefWidth="100" prefHeight="40" buttonType="RAISED" text="COLORED"
+                   style="-fx-text-fill:WHITE;-fx-background-color:#5264AE;-fx-font-size:14px;"/>
         <Button layoutX="388.0" layoutY="195.0" mnemonicParsing="false" onAction="#onClickSignUp" prefHeight="42.0" prefWidth="102.0" text="Registrazione" />
         <Label layoutX="240.0" layoutY="68.0" text="Benvenuto!">
             <font>
