+++ client/src/main/resources/signInLayout.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.layout.Pane?>
 <?import javafx.scene.text.Font?>
 
-<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="300.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.121" xmlns:fx="http://javafx.com/fxml/1" fx:controller="it.cwmp.controller.AuthenticationController">
+<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="300.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.121" xmlns:fx="http://javafx.com/fxml/1" fx:controller="it.cwmp.client.controller.AuthenticationController">
 
     <children>
         <Button layoutX="471.0" layoutY="233.0" mnemonicParsing="false" onAction="#onClickSignIn" prefHeight="42.0" prefWidth="102.0" text="Login" />
