+++ client/src/main/scala/it/cwmp/client/controller/AuthenticationController.scala
@@ -1,4 +1,4 @@
-package it.cwmp.controller
+package it.cwmp.client.controller
 
 import javafx.fxml.FXML
 import javafx.scene.control.Alert.AlertType
