+++ client/src/main/scala/it/cwmp/client/utils/LayoutRes.scala
@@ -1,4 +1,4 @@
-package it.cwmp.utils
+package it.cwmp.client.utils
 
 /**
   * Utility class that keeps together all the file names of the FXML layout files.
