+++ client/src/main/scala/it/cwmp/client/utils/StringRes.scala
@@ -1,4 +1,4 @@
-package it.cwmp.utils
+package it.cwmp.client.utils
 
 /**
   * Utility class that keeps together all the strings that compose the interface.
