+++ client/src/main/scala/it/cwmp/client/utils/ViewUtils.scala
@@ -1,0 +1,7 @@
+package it.cwmp.client.utils
+
+object ViewUtils {
+
+
+
+}
