+++ client/src/main/scala/it/cwmp/client/view/SignInView.scala
@@ -1,7 +1,7 @@
-package it.cwmp.view
+package it.cwmp.client.view
 
-import it.cwmp.controller.{AuthenticationController, OpeningController}
-import it.cwmp.utils.{LayoutRes, StringRes}
+import it.cwmp.client.controller.{AuthenticationController, OpeningController}
+import it.cwmp.client.utils.{LayoutRes, StringRes}
 import javafx.application.Platform
 import javafx.fxml.FXMLLoader
 import javafx.scene.Scene
