+++ client/src/main/scala/it/cwmp/client/view/View.scala
@@ -1,4 +1,4 @@
-package it.cwmp.view
+package it.cwmp.client.view
 
 /**
   * Common trait to all view classes.
