--- client/src/main/scala/it/cwmp/utils/ViewUtils.scala
@@ -1,7 +1,0 @@
-package it.cwmp.utils
-
-object ViewUtils {
-
-
-
-}
