+++ server/authentication/src/main/scala/it/cwmp/authentication/AuthenticationServiceVerticle.scala
@@ -7,7 +7,7 @@ import io.vertx.scala.ext.web.{Router, RoutingContext}
 import io.vertx.core.json.JsonObject
 import io.vertx.scala.ext.jdbc.JDBCClient
 import it.cwmp.storage.StorageAsync
-import it.cwmp.utils.HttpUtils
+import it.cwmp.client.utils.HttpUtils
 
 import scala.concurrent.Future
 import scala.util.{Failure, Success}
