+++ server/rooms/src/test/scala/it/cwmp/room/RoomsServiceTest.scala
@@ -4,7 +4,7 @@ import io.netty.handler.codec.http.HttpHeaderNames
 import io.vertx.scala.ext.web.client.{WebClient, WebClientOptions}
 import it.cwmp.authentication.AuthenticationServiceVerticle
 import it.cwmp.testing.VerticleTesting
-import it.cwmp.utils.HttpUtils
+import it.cwmp.client.utils.HttpUtils
 import org.scalatest.Matchers
 
 import scala.concurrent.{Future, Promise}
